requirements.txt
absl-py==2.1.0
asgiref==3.8.1
astunparse==1.6.3
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
cloudpickle==3.0.0
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
Django==5.0.4
djangorestframework==3.15.1
flatbuffers==24.3.25
fonttools==4.51.0
gast==0.5.4
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
grpcio==1.62.2
h5py==3.11.0
idna==3.7
imageio==2.34.0
joblib==1.4.0
keras==2.15.0
kiwisolver==1.4.5
lazy_loader==0.4
libclang==18.1.1
llvmlite==0.42.0
Markdown==3.6
MarkupSafe==2.1.5
matplotlib==3.7.4
ml-dtypes==0.2.0
networkx==3.3
numba==0.59.1
numpy==1.24.4
oauthlib==3.2.2
opencv-python==4.9.0.80
opt-einsum==3.3.0
packaging==24.0
pandas==2.2.2
pillow==10.3.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
scikit-image==0.22.0
scikit-learn==1.4.2
scipy==1.13.0
seaborn==0.12.2
shap==0.45.0
six==1.16.0
slicer==0.0.7
sqlparse==0.5.0
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-intel==2.15.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
threadpoolctl==3.5.0
tifffile==2024.4.18
tqdm==4.66.2
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
Werkzeug==3.0.2
wrapt==1.14.1
